+++ python/bullmq/utils.py
@@ -1,0 +1,4 @@
+import semver
+
+def isRedisVersionLowerThan(current_version, minimum_version):
+    return semver.compare(current_version, minimum_version) == -1